// apps/kanban-web/src/lib/generated/models/WebsocketEventType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebsocketEventType = 'project.created' | 'project.deleted' | 'card.created' | 'card.branch.updated' | 'card.moved' | 'card.commented' | 'card.updated' | 'card.todo.added' | 'card.todo.updated' | 'card.todo.deleted' | 'card.acceptance.added' | 'card.acceptance.updated' | 'card.acceptance.deleted' | 'card.deleted_soft' | 'card.deleted_hard' | 'resync.required';